+++ app/Http/Controllers/IssuesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Issue;
 use App\Area;
+use App\Contact;
 use App\Task;
 use App\User;
 use App\Priority;
@@ -15,10 +16,12 @@ use Illuminate\Support\Facades\Auth;
 use App\Mail\IssueCreated;
 use Illuminate\Support\Facades\Mail;
 use App\Events\Issues\NewIssue;
-use App\Events\Issues\IssueReopened;
 use App\Events\Issues\UpdatedIssue;
+use App\Events\Issues\IssueClosed;
+use App\Events\Issues\IssueReopened;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Cache;
+use Illuminate\Support\Arr;
 class IssuesController extends Controller
 {
 	
@@ -113,7 +116,7 @@ class IssuesController extends Controller
 				'E_post' => $item->customerMail,
 				'Telefon' => $item->customerTel,
 				'Skapad_av' => $item->userCreate->fullName(),
-				'Senaste_kommentar' => $item->latestComment['comment_internal'],
+				'Senaste_kommentar' => $item->latestComment['comment'],
 
 				'_rowVariant' => $rowVariant,
             ];
@@ -188,11 +191,13 @@ class IssuesController extends Controller
      */
     public function show(Issue $issue)
     {
-		$Comments = IssueComment::
+		$comments = IssueComment::
 			where('issue_id',$issue->id)
 			->hasComments()
+			->orderBy('checkin', 'desc')
 			->get();
-				
+			// external contact means outside Enterprise, ie customers
+		$contacts = Contact::where('external', 0)->get();		
         $areas = Area::all();
         $tasks = Task::all();
 		$users = User::where('active', 1)->get();
@@ -209,9 +214,16 @@ class IssuesController extends Controller
 		$issue->refresh();
 		$new_comment = check_out_issue($issue);
 		$issue->refresh();
+		$selected = $contacts->map(function ( $contact ) {
+			return [
+				'value' => $contact->id,
+				'text' => $contact->name,
+			];
+		});
+
 		return view('issues.show')->with([
 			'issue' => $issue, 
-			'comments' => $Comments,
+			'comments' => $comments,
 			'areas' => $areas, 
 			'tasks' => $tasks, 
 			'users' => $users,
@@ -219,6 +231,7 @@ class IssuesController extends Controller
 			'followers' => $followers,
 			'follow' => $follow,
 			'new_comment' => $new_comment,
+			'contacts' => $selected,
 			]);
     }
 
@@ -294,12 +307,26 @@ class IssuesController extends Controller
 		return redirect()->back();
 	}
 	
+	public function close($id)
+	{
+
+		// $validatedData['timeClosed'] = date('Y-m-d H:i:s');
+		$issue = Issue::find($id);
+		$issue->update([
+		'timeClosed' => date('Y-m-d H:i:s')
+		]);
+		event(new IssueClosed($issue));
+		event(new UpdatedIssue($issue, $type='comment',[]));
+		return redirect('/issues');
+		
+	}
+
 	public function reopen($id)
 	{
 		$issue = Issue::find($id);
 		Issue::whereId($id)->update(['timeClosed' => null]);
 		event(new IssueReopened($issue));
-		event(new UpdatedIssue($issue, $type='comment'), []);
+		event(new UpdatedIssue($issue, $type='comment', []));
 		return redirect()->back();
 	}