+++ app/Listeners/Issues/GenerateReopenedComment.php
@@ -32,9 +32,9 @@ class GenerateReopenedComment
 		$new_comment = new IssueComment;
 		$new_comment->issue_id = $event->issue->id;
 		$new_comment->user_id = Auth::id();
-		$new_comment->comment_internal = 'Ärendet öppnat igen';
-		$new_comment->checkout = date('Y-m-d H:i',strtotime(now()));
-		$new_comment->checkin = date('Y-m-d H:i',strtotime(now()));
+		$new_comment->comment = 'Ärendet öppnat igen';
+		$new_comment->checkout = date('Y-m-d H:i:s',strtotime(now()));
+		$new_comment->checkin = date('Y-m-d H:i:s',strtotime(now()));
 		$new_comment->Save();
     }
 }