+++ resources/js/components/IssueComment.vue
@@ -1,0 +1,24 @@
+<template>
+<div class="my-1">
+    <small>
+        <strong>
+            <slot name="date"></slot>
+            <slot name="type"></slot>
+            <slot name="from"></slot>
+            <slot name="to"></slot>
+        </strong>
+    </small>
+    <slot></slot>
+</div>
+</template>
+
+<script>
+export default {
+    data() {
+        return {
+            icon: 'info',
+            to: '',
+        }
+    },
+};
+</script>